requirements.txt
# This file can be used to create an environment using:
# conda create --name <env> --file <filename>
numpy=1.18.1
pytorch=1.4.0
torchvision=0.5.0
tqdm=4.44.1
numpy=1.18.1
opencv=3.4.2
pillow=7.0.0
imgaug=0.4.0
tensorboard=1.15.0
